requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
matplotlib
numpy
opencv-python
